requirements.txt
# Core API and Processing
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Meta Llama and LLM Integration
transformers==4.36.0
torch==2.1.0
tokenizers==0.15.0
accelerate==0.24.1

# Cerebras API
cerebras-cloud-sdk==1.0.0

# Vector Database
chromadb==0.4.15
sentence-transformers==2.3.1
huggingface-hub==0.20.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-magic==0.4.27

# Image Processing
Pillow==10.1.0
ftfy==6.1.1
regex==2023.10.3
tqdm==4.66.1
opencv-python==4.8.1.78
easyocr==1.7.0
pytesseract==0.3.10

# Audio Processing
openai-whisper==20231117
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Embeddings and ML
numpy==1.24.3
scikit-learn==1.3.2
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.0
asyncio==3.4.3
typing-extensions==4.8.0
httpx==0.25.2

# Frontend
streamlit==1.28.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1